# gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Gravitee.io Portal Rest API
  description: API dedicated to the devportal part of Gravitee
  termsOfService: terms
  contact:
    name: The Gravitee Team
    url: https://gravitee.io
    email: contact@graviteesource.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "3.0.0"
servers:
  - url: http://localhost:8083/portal/environments/{envId}
    description: The portal API for a given environment
    variables:
      envId:
        default: DEFAULT
        description: ID of the environment
...
security:
  - BasicAuth: []
  - CookieAuth: []
paths:
...
  /users/{userId}/avatar:
    parameters:
      - $ref: '#/components/parameters/userIdParam'
    get:
      tags:
        - Users
      summary: Retrieve a user's avatar
      description: |
        Retrieve a user's avatar.
      operationId: getUserAvatar
      responses:
        200:
          $ref: '#/components/responses/PictureSuccess'
        500:
          $ref: '#/components/responses/InternalServerError'
  /user:
    get:
      tags:
        - User
      summary: Get the authenticated user
      description: |
        Get information about the authenticated user.
      operationId: getCurrentUser
      responses:
        200:
          description: The current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        500:
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - User
      requestBody:
        description: Use to update a user.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      summary: Modify current user information.
      description: |
        Modify current user information.

        Only the current user can modify his/her information.
      operationId: updateCurrentUser
      responses:
        200:
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Input must not be null.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Not allowed to modify another user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
  /user/notifications:
    get:
      tags:
        - User
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
      summary: Retrieve user's notifications
      description: |
        Retrieve current user's notifications.
      operationId: getCurrentUserNotifications
      responses:
        200:
          description: List of notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalNotificationsResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - User
      summary: Delete all notifications of the current user
      description: |
        Delete all notifications of the current user.
      operationId: deleteAllCurrentUserNotifications
      responses:
        204:
          description: No-Content
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /user/avatar:
    get:
      tags:
        - User
      summary: Retrieve user's avatar
      description: |
        Retrieve user's avatar.
      operationId: getCurrentUserAvatar
      responses:
        200:
          $ref: '#/components/responses/PictureSuccess'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  parameters:
...
    userIdParam:
      name: userId
      in: path
      required: true
      description: Id of a user
      schema:
        type: string
...
    pageNumberParam:
      name: page
      in: query
      required: false
      description: The page number for pagination.
      schema:
        type: integer
        default: 1
    pageSizeParam:
      name: size
      in: query
      required: false
      description: |
        The number of items per page for pagination.
        If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
        If the size is -1, the response contains all datas.
      schema:
        type: integer
        default: 10

    #Apis filtering
...
  schemas:
...
    ErrorResponse:
      properties:
        errors:
          description: List of errors
          type: array
          items:
            $ref: '#/components/schemas/Error'
...
    PortalNotificationsResponse:
      properties:
        data:
          description: List of portal notifications.
          type: array
          items:
            $ref: '#/components/schemas/PortalNotification'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    Error:
      properties:
        status:
          description: Status HTTP of the error
          type: string
        message:
          description: Message of the error
          type: string
        code:
          description: Technical code of the error
          type: string
        parameters:
          description: Parameters
          type: object
          additionalProperties:
            type: string
...
    PortalNotification:
      properties:
        id:
          description: Unique identifier of a portal notification.
          type: string
        title:
          description: Title of the portal notification.
          type: string
        message:
          description: Content of the notification.
          type: string
        created_at:
          description: Creation date and time of the notification.
          type: string
          format: date-time
...
    User:
      properties:
        id:
          description: Unique identifier of a user.
          type: string
        reference:
          description: Unique reference if user comes from external source. Use for search only.
          type: string
        first_name:
          type: string
        last_name:
          type: string
        display_name:
          type: string
        email:
          type: string
        editable_profile:
          description: True if the user can edit the MyAccount information
          type: boolean
          default: false
        permissions:
          description: Map of Permissions.
          type: object
          properties:
            INSTANCE:
              type: array
              items:
                type: string
            GROUP:
              type: array
              items:
                type: string
            TAG:
              type: array
              items:
                type: string
            TENANT:
              type: array
              items:
                type: string
            API:
              type: array
              items:
                type: string
            APPLICATION:
              type: array
              items:
                type: string
            PLATFORM:
              type: array
              items:
                type: string
            AUDIT:
              type: array
              items:
                type: string
            NOTIFICATION:
              type: array
              items:
                type: string
            MESSAGE:
              type: array
              items:
                type: string
            DICTIONARY:
              type: array
              items:
                type: string
            ALERT:
              type: array
              items:
                type: string
            ENTRYPOINT:
              type: array
              items:
                type: string
            SETTINGS:
              type: array
              items:
                type: string
            QUALITY_RULE:
              type: array
              items:
                type: string
            DASHBOARD:
              type: array
              items:
                type: string
            METADATA:
              type: array
              items:
                type: string
            DOCUMENTATION:
              type: array
              items:
                type: string
            CATEGORY:
              type: array
              items:
                type: string
            TOP_APIS:
              type: array
              items:
                type: string
            API_HEADER:
              type: array
              items:
                type: string
            IDENTITY_PROVIDER:
              type: array
              items:
                type: string
            CLIENT_REGISTRATION_PROVIDER:
              type: array
              items:
                type: string
            THEME:
              type: array
              items:
                type: string
        config:
          $ref: '#/components/schemas/UserConfig'
        _links:
          $ref: '#/components/schemas/UserLinks'
    UserConfig:
      properties:
        management_url:
          description: The URL of the Gravitee management UI
          type: string
...
    UserInput:
      required:
        - id
      properties:
        id:
          description: Unique identifier of a user.
          type: string
        avatar:
          description: Avatar of the user.
          type: string
          writeOnly: true
        first_name:
          description: Firstname of the user.
          type: string
          writeOnly: true
        last_name:
          description: Lastname of the user.
          type: string
          writeOnly: true
        email:
          description: Email of the user.
          type: string
          writeOnly: true
...
    MetadataMap:
      description: Map of Map of Object
      type: object
      additionalProperties:
        type: object
        additionalProperties:
          type: object
...
    Links:
      properties:
        self:
          type: string
        first:
          type: string
        last:
          type: string
        prev:
          type: string
        next:
          type: string
...
    UserLinks:
      properties:
        self:
          type: string
        avatar:
          type: string
        notifications:
          type: string
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PictureSuccess:
      description: A picture
      content:
        'image/*':
          schema:
            type: string
            format: binary
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    CookieAuth:
      type: apiKey
      in: cookie
      name: Auth-Graviteeio-APIM